backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings
faiss-cpu
sentence-transformers
PyMuPDF
numpy
